remove joint from ownedbadge

+++ Back/API/src/main/java/com/viveris/api/service/OwnedBadgeService.java
@@ -26,7 +26,7 @@ public class OwnedBadgeService {
 		return ownedBadgeRepository.findAll();
 	}
 	
-	public Iterable<OwnedBadge> getOwnedBadgesByUserId(Integer user_id) {
+	public Iterable<OwnedBadge> getOwnedBadgesByUserId(Long user_id) {
 		return ownedBadgeRepository.findByUser_Uid(user_id);
 	}
